pyproject.toml
[tool.pytest.ini_options]
pythonpath = [
    ".", "app",
]
asyncio_mode = "auto"

[tool.poetry]
name = "app"
version = "0.0.1"
description = "Packages for application"
authors = ["agent-yandex <fox.online12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

fastapi = "^0.111.0"
python-multipart = "^0.0.9"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
pyjwt = "^2.8.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
passlib = "^1.7.4"
pydantic-marshals = "^0.3.13"
greenlet = "^3.0.3"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
faker = "^25.8.0"
rstr = "^3.2.2"
black = "^24.4.2"
flake8 = "^7.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
